feat: add NavBar
- uses AppBar and renders on App.txt only for now
